ifupdown: basic support for bridge_ports
This patch adds basic support for setting up bridges with ifupdown.
The implementation is limited to listing actual interfaces in the
bridge_ports attribute of a stanza in /etc/network/interace:

    iface br0 inet static
        address 192.168.1.1
	netmask 255.255.255.0
	bridge_ports lan0 lan1

The patch hooks on to the manual method callbacks, adding the bridge
interface if it's missing and removing it when taking it down.  Like
the bridge-utils-interfaces extension of ifupdown in Debian does.

The patch takes care to support ifconfig, ip link/addr and the brctl
tool depending on the BusyBox config.

Other bridge_ attributes, as well as regexp ifname matching for the
bridge_ports attribute can be added later.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>