ntpd: always create PID file
Regardless if the process runs in foregrund or background, as a courtesy
to users, we should always create the PID file.  Modern init systems run
daemons in the foreground to monitor them, and some use the creation of
PID file as a sign to start any other process that may depend on it.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>